refactor: group operations to the same folder
