+++ src/index.ts
@@ -10,10 +10,10 @@ import {
     base64ToHex,
     hexToBase64
 } from "./types";
-import { randomBytes } from "./random";
-import { sha256, sha512 } from "./sha2";
-import { kdf } from "./kdf";
-import { ppf } from "./ppf";
-import { hmac } from "./hmac";
-import { generateEd25519KeyPair, sign, verify } from "./ed25519";
-import { generateRSAKeyPair, exportAsPem, rsaEncrypt, rsaDecrypt } from "./rsa";
+import { randomBytes } from "./operations/random";
+import { sha256, sha512 } from "./operations/sha2";
+import { kdf } from "./operations/kdf";
+import { ppf } from "./operations/ppf";
+import { hmac } from "./operations/hmac";
+import { generateEd25519KeyPair, sign, verify } from "./operations/ed25519";
+import { generateRSAKeyPair, exportAsPem, rsaEncrypt, rsaDecrypt } from "./operations/rsa";
