+++ tests/operations/sha2.spec.ts
@@ -1,5 +1,5 @@
-import { ByteArray, bytesToHex } from "../src/types";
-import { sha256, sha512 } from "../src/sha2";
+import { ByteArray, bytesToHex } from "../../src/types";
+import { sha256, sha512 } from "../../src/operations/sha2";
 
 describe("SHA2", () => {
     const input: string = "hash me!";
